time_entry/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-environ==0.4.5
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
google==3.0.0
google-cloud==0.34.0
idna==2.8
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mysql-connector-python==8.0.22
packaging==20.3
paramiko==2.7.2
pep517==0.8.2
Pillow==8.0.1
pipenv==2020.11.15
progress==1.5
protobuf==3.13.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
urllib3==1.25.8
virtualenv==20.2.2
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
